removing BoxedLink and its story
