--- src/components/display/fewlines/BoxedLink/BoxedLink.stories.tsx
@@ -1,18 +1,0 @@
-import React from "react";
-
-import { Container } from "../Container";
-import { BoxedLink } from "./BoxedLink";
-
-export default { title: "components/Boxed Link", component: BoxedLink };
-
-export const StandardBoxedLink = (): JSX.Element => {
-  return (
-    <Container>
-      <BoxedLink
-        value={"emailaddress@mail.test"}
-        primary={true}
-        status="validated"
-      />
-    </Container>
-  );
-};
