+++ tests/pages/LoginsOverviewPage.test.tsx
@@ -4,13 +4,13 @@ import React from "react";
 
 import { IdentityTypes } from "@lib/@types";
 import { SortedIdentities } from "@src/@types/SortedIdentities";
-import { BoxedLink } from "@src/components/display/fewlines/BoxedLink/BoxedLink";
 import { H1 } from "@src/components/display/fewlines/H1/H1";
 import { H2 } from "@src/components/display/fewlines/H2/H2";
 import { Value } from "@src/components/display/fewlines/LoginsOverview/LoginsOverview";
+import { BoxedLink } from "@src/components/display/fewlines/LoginsOverview/LoginsOverview";
 import { ShowMoreButton } from "@src/components/display/fewlines/ShowMoreButton/ShowMoreButton";
 import { AccountApp } from "@src/pages/_app";
-import LoginsOverviewPage from "@src/pages/account/logins/index";
+import LoginsOverviewPage from "@src/pages/account/logins";
 
 jest.mock("@src/config", () => {
   return {
